refactor: 응답 필드 수정
