+++ src/pages/Write/components/DetailFormCard.tsx
@@ -10,7 +10,7 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
-import { useFetchThread } from "@/pages/Home/apis/fetchThread";
+import { useFetchThread } from "@/pages/Home/apis/useFetchThread";
 import { useUpdateThread } from "@/pages/Write/apis/useUpdateThread";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
